requirements.txt
Flask==1.0.2
numpy==1.16.3
tqdm==4.39.0
werkzeug==0.16.0
ipdb==0.12.2
seqeval==0.0.3
requests==2.22.0
pandas==1.0.5
tensorflow-gpu==1.12.0
jieba
zhon
xlrd